requirements.txt
groq>=0.9.0
pandas>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
